refactor: rewrite table view components and state management with Svelte 5 runes
